src/consulta_bases_de_distribuicao_e_trr_autorizados/requirements.txt
pandas
pandas_gbq
google-cloud-bigquery
google-cloud-storage
requests
beautifulsoup4
lxml
python-
openpyxl
